requirements.txt
async-generator==1.10
attrs==22.2.0
black @ file:///C:/b/abs_620t6ndje8/croot/black_1680737261963/work
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama @ file:///C:/b/abs_a9ozq0l032/croot/colorama_1672387194846/work
contourpy @ file:///C:/b/abs_d5rpy288vc/croots/recipe/contourpy_1663827418189/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
exceptiongroup==1.0.4
fonttools==4.25.0
h11==0.14.0
idna==3.4
kiwisolver @ file:///C:/ci/kiwisolver_1653292359073/work
libcst==0.4.9
matplotlib @ file:///C:/b/abs_03z_dj2gty/croot/matplotlib-suite_1670466145509/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
MonkeyType==22.2.0
munkres==1.1.4
mypy-extensions==0.4.3
mysql-connector-python==8.0.33
numexpr @ file:///C:/b/abs_a7kbak88hk/croot/numexpr_1668713882979/work
numpy==1.24.0
outcome==1.2.0
packaging==23.1
pandas @ file:///C:/b/abs_c6fuezktfm/croot/pandas_1670425103552/work
pathspec==0.11.1
Pillow==9.3.0
platformdirs==3.5.0
protobuf==3.20.3
pycodestyle @ file:///C:/b/abs_d77nxvklcq/croot/pycodestyle_1674267231034/work
pycparser==2.21
pyparsing @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_7f_7lba6rl/croots/recipe/pyparsing_1661452540662/work
PySocks==1.7.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
seaborn @ file:///C:/b/abs_64ypwhwrqe/croot/seaborn_1669625732106/work
selenium==4.7.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.3.0
sortedcontainers==2.4.0
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli @ file:///C:/Windows/TEMP/abs_ac109f85-a7b3-4b4d-bcfd-52622eceddf0hy332ojo/croots/recipe/tomli_1657175513137/work
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.9.2
typing-inspect==0.8.0
typing_extensions==4.4.0
undetected-chromedriver==3.1.7
Unidecode==1.3.6
urllib3==1.26.13
webdriver-manager==3.8.6
websockets==10.4
wincertstore==0.2
wsproto==1.2.0
